requirements.txt
matplotlib
numpy
pynetbox
flask
python-dotenv
